Create MAC learning contoller
The difference in RTT can be attributed to the different network paths that the ping packets take to reach h2 and h5 from h1. In a network controlled by a MAC learning controller, the controller learns the mapping between MAC addresses and ports and instructs the switch which port to forward the packet on. This means that the network path between two hosts can be more efficient compared to a hub-like controller, as the traffic is not flooded to all ports. However, the path can still be longer for some host pairs compared to others, leading to higher network latency for those host pairs, as observed in the case of h1 and h5.